src/requirements.txt
djangorestframework
psycopg2-binary
drf-yasg
djoser
djangorestframework_simplejwt
gunicorn
django-filter
